+++ inst/shiny/global.R
@@ -1,6 +1,7 @@
 shinyscholar::check_suggests()
 
 library(glue)
+library(shinyscholar)
 
 future::plan(future::multisession)
 
@@ -26,6 +27,8 @@ base_module_configs <- c(
   "modules/select_user.yml",
   "modules/plot_hist.yml",
   "modules/plot_scatter.yml",
+  "modules/plot_auto.yml",
+  "modules/plot_semi.yml",
   "modules/rep_markdown.yml",
   "modules/rep_renv.yml",
   "modules/rep_refPackages.yml",