+++ inst/shiny/modules/rep_markdown.R
@@ -5,15 +5,14 @@ rep_markdown_module_ui <- function(id) {
     strong("Select download file type"),
     selectInput(ns("rmdFileType"), label = "",
                 choices = c("Rmd" = ".Rmd", "PDF" = ".pdf", "HTML" = ".html", "Word" = ".docx")),
-    downloadButton(ns("dlRMD"), 'Download Session Code')
+    downloadButton(ns("download"), 'Download Session Code')
   )
 }
 
 rep_markdown_module_server <- function(id, common, parent_session, map, COMPONENT_MODULES) {
   moduleServer(id, function(input, output, session) {
 
-    # handler for R Markdown download
-    output$dlRMD <- downloadHandler(
+    output$download <- downloadHandler(
       filename = function() {
         paste0("shinyscholar-session-", Sys.Date(), input$rmdFileType)
       },
@@ -73,6 +72,33 @@ rep_markdown_module_server <- function(id, common, parent_session, map, COMPONEN
         result_file <- tempfile(pattern = "result_", fileext = input$rmdFileType)
         if (input$rmdFileType == ".Rmd") {
           combined_rmd <- gsub("``` r", "```{r}", combined_md)
+          combined_rmd <- unlist(strsplit(combined_rmd , "\n"))
+
+          # add title section
+          header <- c("---", paste0("title: ", combined_rmd[1]), "---")
+          combined_rmd <- append(combined_rmd, header, after = 0)
+          combined_rmd <- combined_rmd[-c(4, 5)]
+
+          # convert chunk control lines
+          chunk_control_lines <- grep("\\{r,", combined_rmd)
+          if (length(chunk_control_lines) > 0){
+            chunk_starts <- grep("```\\{r\\}", combined_rmd)
+            chunks_to_remove <- NA
+            for (i in seq_along(chunk_control_lines)) {
+              chunks_to_remove[i] <- min(chunk_starts[chunk_starts > chunk_control_lines[i]])
+            }
+            combined_rmd <- combined_rmd[-chunks_to_remove]
+            combined_rmd <- gsub("\\{r,", "```{r,", combined_rmd)
+          }
+
+          # fix any very long lines
+          long_lines <- which(nchar(combined_rmd) > 4000)
+          for (l in long_lines){
+            split_lines <- strwrap(combined_rmd[l], 4000)
+            combined_rmd <- combined_rmd[-l]
+            combined_rmd <- append(combined_rmd, split_lines, l-1)
+          }
+
           writeLines(combined_rmd, result_file, useBytes = TRUE)
         } else {
           combined_md_file <- tempfile(pattern = "combined_", fileext = ".md")