+++ NEWS.md
@@ -69,3 +69,29 @@ shinyscholar 0.2.5
 - Fix bug in writeLog when `type = warning`
 - Skip tests that download files on Fedora systems
 - Module `run` buttons can be pressed using the Enter key
+
+shinyscholar 0.3.0
+=============
+
+### Bug fixes
+- Fixed `core_code` module when no function exists
+- Break up very long lines in the markdown so that they can be read in again
+- Disabled `register_module()` as it has not been refactored to work for created apps 
+- Made loading process more robust by adding the app name 
+- Made loading fault-tolerant if deprecated `common` objects exist
+
+### Changes
+- Tidied up `library()` calls
+- Removed unnecessary `gargoyle::` and `leaflet::` scoping 
+
+### New features
+- Entries in the logger are now restored on load and made available in testing
+- Added `plot_auto` and `plot_semi` modules to the example app that run automatically and semi-automatically respectively
+- Added ability to control generated markdown chunks using `asis` chunks and example to `plot_hist.Rmd`
+- Added `common$reset()` and `reset_data()` to reset the `common` object and remove all outputs
+- Added an area to the UI for including global options available inside all modules
+- Added an options to `create_template()` and `create_module()` to include a `downloadButton` and `downloadHandler` to modules that is only visible once the module runs successfully
+- Rewrote `create_template()` and `create_module()` so that `create_module()` creates the module function and tests
+- Switched to use `shinyWidgets::radioGroupButtons()` for module selection menu
+- Added a description of the package structure to the README
+